move BalanceView to components/display
