+++ src/components/display/BalanceView.js
