+++ src/screens/Assets/deposits/DepositsTab.js
@@ -27,7 +27,7 @@ import styled from 'styled-components/native';
 import { useTranslationWithPrefix } from 'translations/translate';
 
 // Components
-import BalanceView from 'components/BalanceView';
+import BalanceView from 'components/display/BalanceView';
 import ChainListHeader from 'components/lists/ChainListHeader';
 import ChainListFooter from 'components/lists/ChainListFooter';
 import FiatChangeView from 'components/display/FiatChangeView';
