+++ src/screens/KeyBasedAssetTransfer/KeyBasedAssetTransferConfirm.js
@@ -27,7 +27,7 @@ import { useTranslationWithPrefix } from 'translations/translate';
 
 // Components
 import * as Table from 'components/layout/Table';
-import BalanceView from 'components/BalanceView';
+import BalanceView from 'components/display/BalanceView';
 import Button from 'components/legacy/Button';
 import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
 import FeeTable from 'components/display/FeeTable';
